pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Physical AI Humanoid Robotics Textbook"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "uuid>=1.30",
    "python-multipart>=0.0.6",
    "cryptography>=41.0.8",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.13.1",
    "openai-agents>=0.3.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
